Share and simplify ChatPreviewHandler code in 1.19+ Reduces plugin jar by ~2.4 kB Makes code easier to understand and maintain
