+++ commandapi-nms/commandapi-1.19.1/src/main/java/dev/jorel/commandapi/nms/NMS_1_19_1_R1_ChatPreviewHandler.java
@@ -1,58 +1,28 @@
 package dev.jorel.commandapi.nms;
 
-import java.lang.reflect.Field;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-
+import dev.jorel.commandapi.preprocessor.RequireField;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.PacketSendListener;
+import net.minecraft.network.chat.ChatPreviewCache;
+import net.minecraft.network.chat.ChatPreviewThrottler;
+import net.minecraft.network.protocol.game.ClientboundChatPreviewPacket;
+import net.minecraft.network.protocol.game.ServerboundChatPreviewPacket;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import org.bukkit.Bukkit;
 import org.bukkit.craftbukkit.v1_19_R1.entity.CraftPlayer;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
-import com.mojang.brigadier.ParseResults;
-import com.mojang.brigadier.context.ParsedCommandNode;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
-
-import dev.jorel.commandapi.CommandAPIHandler;
-import dev.jorel.commandapi.arguments.PreviewInfo;
-import dev.jorel.commandapi.exceptions.WrapperCommandSyntaxException;
-import dev.jorel.commandapi.preprocessor.RequireField;
-import dev.jorel.commandapi.wrappers.PreviewableFunction;
-import io.netty.channel.ChannelDuplexHandler;
-import io.netty.channel.ChannelHandlerContext;
-import net.kyori.adventure.text.Component;
-import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
-import net.md_5.bungee.api.chat.BaseComponent;
-import net.md_5.bungee.api.chat.TextComponent;
-import net.md_5.bungee.chat.ComponentSerializer;
-import net.minecraft.commands.CommandSourceStack;
-import net.minecraft.network.Connection;
-import net.minecraft.network.PacketSendListener;
-import net.minecraft.network.chat.ChatPreviewCache;
-import net.minecraft.network.chat.ChatPreviewThrottler;
-import net.minecraft.network.chat.ChatPreviewThrottler.Request;
-import net.minecraft.network.chat.Component.Serializer;
-import net.minecraft.network.protocol.game.ClientboundChatPreviewPacket;
-import net.minecraft.network.protocol.game.ServerboundChatPreviewPacket;
-import net.minecraft.server.network.ServerGamePacketListenerImpl;
+import java.lang.reflect.Field;
+import java.util.concurrent.CompletableFuture;
 
 @RequireField(in = ServerGamePacketListenerImpl.class, name = "chatPreviewCache", ofType = ChatPreviewCache.class)
 @RequireField(in = ServerGamePacketListenerImpl.class, name = "chatPreviewThrottler", ofType = ChatPreviewThrottler.class)
-public class NMS_1_19_1_R1_ChatPreviewHandler extends ChannelDuplexHandler {
-
-	private final NMS<CommandSourceStack> nms;
-	private final Plugin plugin;
-	private final Player player;
-	private final Connection connection;
+public class NMS_1_19_1_R1_ChatPreviewHandler extends NMS_1_19_Common_ChatPreviewHandler {
 	ChatPreviewThrottler throttler;
 
 	public NMS_1_19_1_R1_ChatPreviewHandler(NMS<CommandSourceStack> nms, Plugin plugin, Player player) {
-		this.nms = nms;
-		this.plugin = plugin;
-		this.player = player;
-		this.connection = ((CraftPlayer) player).getHandle().connection.connection;
+		super(nms, plugin, player);
 
 		try {
 			Field f = ServerGamePacketListenerImpl.class.getDeclaredField("M");
@@ -64,121 +34,34 @@ public class NMS_1_19_1_R1_ChatPreviewHandler extends ChannelDuplexHandler {
 	}
 
 	@Override
-	public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-		if (msg instanceof ServerboundChatPreviewPacket chatPreview) {
-			if (!chatPreview.query().isEmpty() && chatPreview.query().charAt(0) == '/') {
-				// We want to run this synchronously, just in case there's some funky async
-				// stuff going on here
+	protected void handleChatPreviewPacket(ServerboundChatPreviewPacket chatPreview) {
+		// We want to run this synchronously, just in case there's some funky async stuff going on here
+		throttler.schedule(() -> {
+			int i = chatPreview.queryId();
+			CompletableFuture<net.minecraft.network.chat.Component> result = new CompletableFuture<>();
 
-				throttler.schedule(new Request() {
+			// get preview
+			Bukkit.getScheduler().runTask(this.plugin, () -> result.complete(chatPreviewQueryToJson(chatPreview.query())));
 
-					@Override
-					public CompletableFuture<?> run() {
-						int i = chatPreview.queryId();
-						String s = chatPreview.query();
-
-						return generateComponentToSend(s).thenAccept(component -> {
-							connection.send(new ClientboundChatPreviewPacket(i, component),
-								PacketSendListener.exceptionallySend(() -> new ClientboundChatPreviewPacket(i, null)));
-						});
-					}
-
-				});
-				return;
-			}
-		}
-
-		// Normal packet handling
-		super.channelRead(ctx, msg);
-	}
-
-	@SuppressWarnings("unchecked")
-	private CompletableFuture<net.minecraft.network.chat.Component> generateComponentToSend(String chatPreviewQuery) {
-		CompletableFuture<net.minecraft.network.chat.Component> result = new CompletableFuture<>();
-		
-		Bukkit.getScheduler().runTask(this.plugin, () -> {
-
-			// Substring 1 because we want to get rid of the leading /
-			final String fullInput = chatPreviewQuery.substring(1);
-			ParseResults<CommandSourceStack> results = nms.getBrigadierDispatcher().parse(fullInput, ((CraftPlayer) player).getHandle().createCommandSourceStack());
-
-			// Generate the path for lookup
-			List<String> path = new ArrayList<>();
-			for (ParsedCommandNode<CommandSourceStack> commandNode : results.getContext().getNodes()) {
-				path.add(commandNode.getNode().getName());
-			}
-			Optional<PreviewableFunction<?>> preview = CommandAPIHandler.getInstance().lookupPreviewable(path);
-			if (preview.isPresent()) {
-				// Calculate the (argument) input and generate the component to send
-				String input = results.getContext().getNodes().get(results.getContext().getNodes().size() - 1).getRange().get(fullInput);
-
-				final String jsonToSend;
-
-				Object component = null;
+			// update player's ChatPreviewCache
+			result.thenAcceptAsync(component -> {
 				try {
-					@SuppressWarnings("rawtypes")
-					final PreviewInfo previewInfo;
-					if (CommandAPIHandler.getInstance().lookupPreviewableLegacyStatus(path)) {
-						BaseComponent[] parsedInput = null;
-						try {
-							parsedInput = nms.getChat(results.getContext().build(fullInput), path.get(path.size() - 1));
-						} catch (CommandSyntaxException e) {
-							throw new WrapperCommandSyntaxException(e);
-						}
-						previewInfo = new PreviewInfo<BaseComponent[]>(this.player, input, chatPreviewQuery, parsedInput);
-					} else {
-						Component parsedInput = null;
-						try {
-							parsedInput = nms.getAdventureChat(results.getContext().build(fullInput), path.get(path.size() - 1));
-						} catch (CommandSyntaxException e) {
-							throw new WrapperCommandSyntaxException(e);
-						}
-						previewInfo = new PreviewInfo<Component>(this.player, input, chatPreviewQuery, parsedInput);
-					}
-
-					component = preview.get().generatePreview(previewInfo);
-				} catch (WrapperCommandSyntaxException e) {
-					component = TextComponent.fromLegacyText(e.getMessage() == null ? "" : e.getMessage());
+					Field f = ServerGamePacketListenerImpl.class.getDeclaredField("L");
+					f.setAccessible(true);
+					ChatPreviewCache c = (ChatPreviewCache) f.get(((CraftPlayer) player).getHandle().connection);
+					c.set(chatPreview.query().substring(1), component);
+				} catch (Exception e) {
+					e.printStackTrace();
 				}
+			});
 
-				if (component != null) {
-					if (component instanceof BaseComponent[] baseComponent) {
-						jsonToSend = ComponentSerializer.toString(baseComponent);
-					} else if (CommandAPIHandler.getInstance().getPaper().isPresent()) {
-						if (component instanceof Component adventureComponent) {
-							jsonToSend = GsonComponentSerializer.gson().serialize(adventureComponent);
-						} else {
-							throw new IllegalArgumentException("Unexpected type returned from chat preview, got: " + component.getClass().getSimpleName());
-						}
-					} else {
-						throw new IllegalArgumentException("Unexpected type returned from chat preview, got: " + component.getClass().getSimpleName());
-					}
-				} else {
-					throw new NullPointerException("Returned value from chat preview was null");
-				}
-
-				if (jsonToSend != null) {
-					result.complete(Serializer.fromJson(jsonToSend));
-				}
-			}
-
-			// It seems happy to return null, so we'll do that
-			result.complete(null);
+			// send ChatPreviewPacket
+			return result.thenAccept(
+				component -> connection.send(
+					new ClientboundChatPreviewPacket(i, component),
+					PacketSendListener.exceptionallySend(() -> new ClientboundChatPreviewPacket(i, null))
+				)
+			);
 		});
-
-		result.thenAcceptAsync(component -> {
-			try {
-				Field f = ServerGamePacketListenerImpl.class.getDeclaredField("L");
-				f.setAccessible(true);
-				ChatPreviewCache c = (ChatPreviewCache) f.get(((CraftPlayer) player).getHandle().connection);
-				c.set(chatPreviewQuery.substring(1), component);
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		});
-
-		return result;
-
 	}
-
 }
